functions/utils/pyproject.toml
[tool.poetry]
name = "cloud-functions-utils"
version = "0.2.2"
description = "Utilities for Google Cloud Functions."
authors = ["Jan-Benedikt Jagusch <jan.jagusch@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-bigquery = "^1.24.0"
google-cloud-error-reporting = "^0.33.0"
google-cloud-pubsub = "^1.4.3"
google-cloud-storage = "^1.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
